+++ src/sensors/differential_pressure.cpp
@@ -32,7 +32,7 @@ bool DifferentialPressureSensor::publish(float diffPressureHpa) {
 
     std_msgs::Float32 msg;
     msg.data = diffPressureHpa * 100;
-    msg.data += STATIC_PRESSURE_NOISE * normalDistribution_(randomGenerator_);
+    msg.data += static_cast<float>(STATIC_PRESSURE_NOISE * normalDistribution_(randomGenerator_));
     publisher_.publish(msg);
 
     nextPubTimeSec_ = crntTimeSec + PERIOD;
